feat: update to a bigger scope quipuswap
